# swagger.yaml
swagger: "2.0"
info:
  description: This is a simple fastify catalogue API.
  version: "1.0.0"
  title: "Fastify Tutorial catalogue API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "oliverpelham@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/"
tags:
- name: "items"
  description: "Everything about catalogue items"
schemes:
- "https"
- "http"
paths:
  /items:
    post:
      tags:
      - "items"
      summary: "Add an individual item to the catalogue"
      description: ""
      operationId: "addItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "item object that needs to be catalogued"
        required: true
        schema:
          $ref: "#/definitions/ItemRequest"
      responses:
        "200":
          description: "successful operation"
    put:
      tags:
      - "items"
      summary: "Update an individual item to the catalogue"
      description: ""
      operationId: "updateItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "item object to be updated"
        required: true
        schema:
          $ref: "#/definitions/ItemRequest"
      responses:
        "200":
          description: "successful operation"
    get:
      tags:
      - "items"
      summary: "View all items catalogued"
      description: ""
      operationId: "viewItems"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ItemResponse"
definitions:
  ItemRequest:
    type: "object"
    required:
    - "itemType"
    - "itemName"
    properties:
      itemType:
        type: "string"
        example: "Book"
      itemName:
        type: "string"
        example: "Charlotte's Web"
  ItemResponse:
    type: "object"
    properties:
      itemType:
        type: "string"
        example: "Book"
      itemName:
        type: "string"
        example: "Charlotte's Web"
      ItemID:
        type: "string"
        example: "129871281941"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"